requirements.txt
Flask==0.10.1
gunicorn==0.17.2
mysql-connector-python==1.1.6